FaceDetectionAPI/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
numpy==1.17.2
Pillow==6.1.0
protobuf==3.9.1
gunicorn==20.0.4
psycopg2==2.8.3
psycopg2-binary==2.8.3
python-magic==0.4.15
torch==1.4.0
torchvision==0.5.0
